# openapi/v2/mgmt/openapi.bundle.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "0.14.0",
    "title": "Management API",
    "contact": {
      "name": "Supaglue",
      "email": "docs@supaglue.com",
      "url": "https://supaglue.com"
    },
    "description": "#### Introduction\n\nWelcome to the Management API documentation. You can use this API to manage customer integrations and connections.\n\n[![Run in Postman](https://run.pstmn.io/button.svg)](https://www.postman.com/supaglue/workspace/supaglue-public/overview)\n\n#### Base API URL\n\n```\nhttps://api.supaglue.io/mgmt/v2\n```\n"
  },
  "servers": [
    {
      "url": "https://api.supaglue.io/mgmt/v2",
      "description": "Supaglue API"
    }
  ],
  "paths": {
...
    "/providers": {
      "get": {
        "operationId": "getProviders",
        "tags": [
          "Providers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List providers",
        "description": "Get a list of providers",
        "responses": {
          "200": {
            "description": "Providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/provider"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProvider",
        "summary": "Create provider",
        "tags": [
          "Providers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create_provider"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/provider"
                }
              }
            }
          }
        }
      }
    },
    "/providers/{provider_id}": {
      "get": {
        "operationId": "getProvider",
        "summary": "Get provider",
        "tags": [
          "Providers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/provider"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateProvider",
        "summary": "Update provider",
        "tags": [
          "Providers"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/update_provider"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/provider"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteProvider",
        "summary": "Delete provider",
        "tags": [
          "Providers"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/provider"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "provider_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/customers/{customer_id}/connections": {
      "get": {
        "operationId": "getConnections",
        "tags": [
          "Connections"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List connections",
        "description": "Get a list of connections",
        "responses": {
          "200": {
            "description": "Connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/connection"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createConnection",
        "summary": "Create a connection",
        "description": "This endpoint is for providers that use API key or access key for authentication. Please use [Embedded Links](https://docs.supaglue.com/platform/managed-auth#oauth-20) for providers that support Oauth2 for authentication.",
        "tags": [
          "Connections"
        ],
        "parameters": [],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "title": "apollo",
                    "properties": {
                      "provider_name": {
                        "type": "string",
                        "enum": [
                          "apollo"
                        ],
                        "example": "apollo"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "api_key"
                        ],
                        "example": "api_key"
                      },
                      "api_key": {
                        "type": "string",
                        "description": "API key for the connected customer.",
                        "example": "e0c5545d-77b8-458d-b4bf-fc701a9cecac"
                      }
                    },
                    "required": [
                      "provider_name",
                      "type",
                      "api_key"
                    ]
                  },
                  {
                    "type": "object",
                    "title": "gong",
                    "properties": {
                      "provider_name": {
                        "type": "string",
                        "enum": [
                          "gong"
                        ],
                        "example": "gong"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "access_key_secret"
                        ],
                        "example": "access_key_secret"
                      },
                      "access_key": {
                        "type": "string",
                        "description": "Access key for the connected customer.",
                        "example": "e0c5545d-77b8-458d-b4bf-fc701a9cecac"
                      },
                      "access_key_secret": {
                        "type": "string",
                        "description": "Access key secret for the connected customer.",
                        "example": "258a80aa-2e9b-40b0-8d7a-28f858a856f3"
                      }
                    },
                    "required": [
                      "provider_name",
                      "type",
                      "access_key",
                      "access_key_secret"
                    ]
                  },
                  {
                    "type": "object",
                    "title": "salesforce",
                    "properties": {
                      "provider_name": {
                        "type": "string",
                        "enum": [
                          "salesforce"
                        ],
                        "example": "salesforce"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "oauth2"
                        ],
                        "example": "oauth2"
                      },
                      "refresh_token": {
                        "type": "string"
                      },
                      "instance_url": {
                        "type": "string"
                      },
                      "login_url": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "provider_name",
                      "type",
                      "refresh_token",
                      "instance_url"
                    ]
                  },
                  {
                    "type": "object",
                    "title": "hubspot",
                    "properties": {
                      "provider_name": {
                        "type": "string",
                        "enum": [
                          "hubspot"
                        ],
                        "example": "hubspot"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "oauth2"
                        ],
                        "example": "oauth2"
                      },
                      "refresh_token": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "provider_name",
                      "type",
                      "refresh_token"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Connection created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/connection"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "customer_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/customers/{customer_id}/connections/_provider_user_id": {
      "get": {
        "operationId": "getProviderUserId",
        "summary": "Get logged in user ID and details",
        "description": "This endpoint returns the ID of the logged in user (this is sometimes known as the owner ID) and details.",
        "tags": [
          "Connections"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "parameters": [
          {
            "name": "provider_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "hubspot",
                "ms_dynamics_365_sales"
              ],
              "example": "hubspot"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UserId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string",
                      "description": "The ID of the logged in user. The ID originates from your customer's third-party Provider."
                    },
                    "raw_details": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "The raw data from the your customer's third-party Provider."
                    }
                  },
                  "example": {
                    "user_id": "70115e3c-2700-4112-b28f-2706e08570e6",
                    "raw_details": {
                      "user_id": "70115e3c-2700-4112-b28f-2706e08570e6",
                      "user": "test@something.com",
                      "scopes": [
                        "contacts",
                        "automation",
                        "oauth"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "customer_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/customers/{customer_id}/connections/{connection_id}": {
      "get": {
        "operationId": "getConnection",
        "summary": "Get connection",
        "tags": [
          "Connections"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "200": {
            "description": "Connection",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/connection"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "string",
                          "description": "The id of the user associated with this connection. This is sometimes also known as the owner."
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteConnection",
        "summary": "Delete connection",
        "tags": [
          "Connections"
        ],
        "security": [
          {
            "x-api-key": []
          }
        ],
        "responses": {
          "204": {
            "description": "An empty body is returned on successful deletion."
          }
        }
      },
      "parameters": [
        {
          "name": "customer_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "connection_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "API key to allow developers to access the API"
      }
    },
    "schemas": {
...
      "provider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "0c4cb91f-e392-46f5-8934-17496371f932"
          },
          "application_id": {
            "type": "string",
            "example": "273e1c75-3b6d-439a-9b8e-c6ac407df92a"
          },
          "category": {
            "$ref": "#/components/schemas/category"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "oauth2",
              "api_key"
            ],
            "example": "oauth2"
          },
          "name": {
            "$ref": "#/components/schemas/provider_name"
          },
          "config": {
            "$ref": "#/components/schemas/create_provider_config"
          },
          "objects": {
            "$ref": "#/components/schemas/objects"
          },
          "entity_mappings": {
            "type": "array",
            "description": "This is for object/field mappings when you data model using [Entities](https://docs.supaglue.com/platform/entities/overview). The object/field mappings (aka [Entity Mappings](https://docs.supaglue.com/platform/entities/overview#entity-mapping)) here are set by you, the developer, and the mappings apply for all your customers.",
            "items": {
              "$ref": "#/components/schemas/provider_entity_mapping"
            }
          }
        },
        "required": [
          "id",
          "application_id",
          "category",
          "auth_type",
          "name"
        ]
      },
...
      "object_field_mapping": {
        "type": "object",
        "properties": {
          "schema_field": {
            "type": "string",
            "example": "apolla_first_name"
          },
          "mapped_field": {
            "type": "string",
            "example": "FirstName"
          }
        },
        "required": [
          "schema_field"
        ]
      },
      "connection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "e888cedf-e9d0-42c5-9485-2d72984faef2"
          },
          "application_id": {
            "type": "string",
            "example": "d8ceb3ff-8b7f-4fa7-b8de-849292f6ca69"
          },
          "customer_id": {
            "type": "string",
            "example": "my-customer-1"
          },
          "provider_id": {
            "type": "string",
            "example": "677fcfca-cf89-4387-a189-71c885be67bc"
          },
          "provider_name": {
            "$ref": "#/components/schemas/provider_name"
          },
          "category": {
            "$ref": "#/components/schemas/category"
          },
          "instance_url": {
            "type": "string",
            "example": "https://app.hubspot.com/contacts/123456",
            "description": "Instance URL for the connected customer."
          },
          "schema_mappings_config": {
            "description": "This is for field mappings when you data model using [Objects](https://docs.supaglue.com/platform/objects/overview#schemas). The field mappings here are set by your customers using the [Field Mappings API](https://docs.supaglue.com/api/v2/mgmt/update-object-field-mappings).",
            "type": "object",
            "properties": {
              "common_objects": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "object": {
                      "type": "string",
                      "example": "Contact__c"
                    },
                    "field_mappings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/object_field_mapping"
                      }
                    }
                  },
                  "required": [
                    "object",
                    "field_mappings"
                  ]
                }
              },
              "standard_objects": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "object": {
                      "type": "string",
                      "example": "Contact"
                    },
                    "field_mappings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/object_field_mapping"
                      }
                    }
                  },
                  "example": null,
                  "required": [
                    "object",
                    "field_mappings"
                  ]
                }
              }
            }
          },
          "entity_mappings": {
            "type": "array",
            "description": "This is for object/field mappings when you data model using [Entities](https://docs.supaglue.com/platform/entities/overview#entity-mapping). The object/field mappings here are set by your customers using the [Entity Mappings API](https://docs.supaglue.com/api/v2/mgmt/upsert-entity-mapping).",
            "items": {
              "$ref": "#/components/schemas/connection_entity_mapping"
            },
            "example": []
          }
        },
        "required": [
          "id",
          "application_id",
          "customer_id",
          "provider_id",
          "provider_name",
          "category",
          "instance_url"
        ]
      },
      "category": {
        "type": "string",
        "enum": [
          "crm",
          "engagement",
          "enrichment",
          "marketing_automation",
          "no_category"
        ],
        "example": "crm"
      },
...
      "provider_name": {
        "type": "string",
        "enum": [
          "hubspot",
          "salesforce",
          "pipedrive",
          "zendesk_sell",
          "ms_dynamics_365_sales",
          "zoho_crm",
          "capsule",
          "outreach",
          "gong",
          "apollo",
          "salesloft",
          "intercom",
          "linear",
          "clearbit",
          "6sense",
          "marketo",
          "salesforce_marketing_cloud_account_engagement"
        ],
        "example": "hubspot"
      },
      "provider_name_crm": {
        "type": "string",
        "enum": [
          "hubspot",
          "salesforce",
          "pipedrive",
          "zendesk_sell",
          "ms_dynamics_365_sales",
          "zoho_crm",
          "capsule"
        ]
      },
      "provider_name_engagement": {
        "type": "string",
        "enum": [
          "outreach",
          "apollo",
          "salesloft"
        ]
      },
      "provider_name_no_category": {
        "type": "string",
        "enum": [
          "intercom",
          "gong",
          "linear"
        ]
      },
...
      "create_provider": {
        "oneOf": [
          {
            "type": "object",
            "title": "crm",
            "properties": {
              "auth_type": {
                "type": "string",
                "enum": [
                  "oauth2"
                ]
              },
              "config": {
                "$ref": "#/components/schemas/create_provider_config"
              },
              "objects": {
                "$ref": "#/components/schemas/objects"
              },
              "category": {
                "type": "string",
                "enum": [
                  "crm"
                ]
              },
              "name": {
                "$ref": "#/components/schemas/provider_name_crm"
              }
            },
            "required": [
              "auth_type",
              "config",
              "category",
              "name"
            ]
          },
          {
            "type": "object",
            "title": "engagement",
            "properties": {
              "auth_type": {
                "type": "string",
                "enum": [
                  "oauth2"
                ]
              },
              "config": {
                "$ref": "#/components/schemas/create_provider_config"
              },
              "objects": {
                "$ref": "#/components/schemas/objects"
              },
              "category": {
                "type": "string",
                "enum": [
                  "engagement"
                ]
              },
              "name": {
                "$ref": "#/components/schemas/provider_name_engagement"
              }
            },
            "required": [
              "auth_type",
              "config",
              "category",
              "name"
            ]
          },
          {
            "type": "object",
            "title": "no category",
            "properties": {
              "auth_type": {
                "type": "string",
                "enum": [
                  "api_key"
                ]
              },
              "category": {
                "type": "string",
                "enum": [
                  "no_category"
                ]
              },
              "name": {
                "$ref": "#/components/schemas/provider_name_no_category"
              }
            },
            "required": [
              "auth_type"
            ]
          }
        ]
      },
      "update_provider": {
        "oneOf": [
          {
            "type": "object",
            "title": "crm",
            "properties": {
              "auth_type": {
                "type": "string",
                "enum": [
                  "oauth2"
                ]
              },
              "config": {
                "$ref": "#/components/schemas/update_provider_config"
              },
              "objects": {
                "$ref": "#/components/schemas/objects"
              },
              "entity_mappings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/provider_entity_mapping"
                }
              },
              "category": {
                "type": "string",
                "enum": [
                  "crm"
                ]
              },
              "name": {
                "$ref": "#/components/schemas/provider_name_crm"
              }
            },
            "required": [
              "auth_type",
              "config",
              "category",
              "name"
            ]
          },
          {
            "type": "object",
            "title": "engagement",
            "properties": {
              "auth_type": {
                "type": "string",
                "enum": [
                  "oauth2"
                ]
              },
              "config": {
                "$ref": "#/components/schemas/update_provider_config"
              },
              "objects": {
                "$ref": "#/components/schemas/objects"
              },
              "entity_mappings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/provider_entity_mapping"
                }
              },
              "category": {
                "type": "string",
                "enum": [
                  "engagement"
                ]
              },
              "name": {
                "$ref": "#/components/schemas/provider_name_engagement"
              }
            },
            "required": [
              "auth_type",
              "config",
              "category",
              "name"
            ]
          },
          {
            "type": "object",
            "title": "no category",
            "properties": {
              "auth_type": {
                "type": "string",
                "enum": [
                  "oauth2"
                ]
              },
              "config": {
                "$ref": "#/components/schemas/update_provider_config"
              },
              "objects": {
                "$ref": "#/components/schemas/objects"
              },
              "entity_mappings": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/provider_entity_mapping"
                }
              },
              "category": {
                "type": "string",
                "enum": [
                  "no_category"
                ]
              },
              "name": {
                "$ref": "#/components/schemas/provider_name_no_category"
              }
            },
            "required": [
              "auth_type",
              "config",
              "category",
              "name"
            ]
          }
        ]
      },
...
      "provider_entity_mapping": {
        "type": "object",
        "properties": {
          "entity_id": {
            "type": "string",
            "example": "apolla_contact"
          },
          "object": {
            "$ref": "#/components/schemas/standard_object"
          },
          "field_mappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/entity_field_mapping"
            }
          }
        },
        "required": [
          "entity_id"
        ]
      },
...
    },
...
  },
...
}